add option to specify the path to the static directory
It is a pain to have to start the collector/evaluator from the cube bin directory.
This is especially true when you have a custom config.  This small change allows you
to define a "static" option as part of your config to define the path to the cube
static directory.